Cache the default working directory within a session
DiskStore>>defaultWorkingDirectory currently:

1. Calls a primitive to get the image path
2. Converts the resulting string to a byte array
3. Converts the byte array back to a string with ZnCharacterEncoder
4. Converts the string to a path and gets the parent

every time it is called.

As an example of how often it can be called: Using Iceberg to clone an
empty repository, add a small package and synchronise the repository
called #defaultWorkingDirectory around 500 times.  It is called so many
times because FileSystem>>resolve: leads to FileSystem>>resolvePath:
which resolves the supplied path against the working directory, and
#resolve: is called from many places.  (The number of calls to
#defaultWorkingDirectory can vary significantly based on the
environment, but you get the idea).

The value of defaultWorkingDirectory only changes on startup and when the image is saved to a different location.

Caching defaultWorkingDirectory in a class variable and managing the
cache through session startup and shutdown gives around a 1,000 times performance improvement in #defaultWorkingDirectory.