Compare absolute paths
FileReference>>= currently performs a simplistic path comparison to
determine if two FileReferences are the same.

This means that an absolute reference and a relative reference to the
same file will be flagged as different files.

The (soon to be supplied) Pull Request modifies FileReference>>= to
make the paths absolute as part of the comparison, causing the two
FileReferences to be considered equal.

As an example, in Pharo6:

| f1 f2 |

f1 := FileLocator workingDirectory / 'pharo-local'.
f2 := f1 asAbsolute.
f1 = f2
" false"

After this PR, the above answers true.