Yield the CPU longer
The work of the idle process is to run when no other smalltalk process is
runnable. And the process will use a primitive to yield execution for a given
amount of microseconds. The VM will either pick the closest delay to expire
or the timeout provided and yield execution. It will resume earlier in case of
I/O.

Pick 50ms as this is the time WorldState will update in server mode. In the
future we can increase it to minutes.