This commit responds to Issue 20471.
It adds character classification methods on Unicode class for the 29 Unicode categories, and
testing methods (isMathSymbol, etc)  in Character instances that delegate  to
these methods.  Because CharacterSets are pluggable, this commit also implements
the same 29 character classification methods on AsciiCharset class and Latin1 class.   The
methods on AsciiCharset class are self-contained, but those on Latin1 class delegate to Unicode,
as did the existing Latin1 class>>#isLetter method.