Let round:  result be correctly rounded
This is issue:
https://pharo.fogbugz.com/f/cases/15473/Float-round-to-n-decimal-places-is-not-in-agreement-with-printShowingDecimalPlaces

It also fixes:
https://pharo.fogbugz.com/f/cases/15472/Fraction-rounding-to-n-decimal-places-is-inexact
https://pharo.fogbugz.com/f/cases/15471/Can-t-round-Float-fmax-to-2-decimal-places

Note: the new implementation does round exact tie away from zero exactly like #round do.
It is also in agreement with printShowingDecimalPlaces:

If you wish to roundToNearestTieToEven, then please open a new issue
(but it will be better to also modify printShowingDecimalPlaces: then)

The new implementation is a bit slower but it is correct.

It was proposed to deprecate round: in

I’m not so sure - round: can be useful for monetary apps and is available in other languages (Python, etc…)
banker rounding (tie to even) might be a better option though.

Last, I’ve replaced the `< expr: foo result: bar>` annotation with `“ foo >>> bar “`
The reason is that the expr:result: form only works for a literal result which is SEVERLY limited.