Fix the sign of FloatNegativeZero
Like any other zero, the sign of Float negativeZero should be zero…
Introduce signBit to differentiate Float negativeZero from Float zero.

This is issue https://pharo.fogbugz.com/f/cases/19629/0-0-sign-answers-1-should-be-zero

Note that this PR does not implement isSignMinus which is not strictly necessary.
If it’s for ISO 10967 compliance, then there will be other functions missing.
Maybe open a new issue for this one?