Fix potential dangling reference to a released FT_Face
This is issue https://pharo.fogbugz.com/f/cases/20776/Potential-dangling-ref-to-freed-FT_Face

Technical details:

A cairo_font_face could still be referenced from external heap and further used.
Thus we must NOT release the FT_Face until the external reference count drops to 1.

Thus the CairoFontFace is not a good place for retaining a strong reference to the FT2Face, because the reference will vanish as soon as the Smalltalk object is garbage collected, while some external cairo_font_face still point on it.
It's better to have the strong reference in the executor, and avoid garbageCollecting the executor while there are external references to the resource data...

The solution is to use a specific CairoFontFaceExecutor.
If the resource is shared, then we have no other choice than regularly polling the ref count in a new Process.
This process must stop whenever the session change.

Please review carefully and pass some tests.