All users of Context based mirror primitives are now using MirrorPrimitives. MirrorPrimitives includes ModificationForbidden logic with tests
ModificationForbidden subclassed from Error again. 
Because instead running all tests with some readonly problems will spawn debugger. Because generally exceptions do not catched by SUnit.
But I think this ModificationForbidden must be a error. The fact that it can be fixed and resumed do not made it notification or warning. It is similar to MessageNotUnderstood which is resumable too