Integrating new implementation of Traits
Separating all the trait related stuff (Traits, extensions, uses) to its specific packages.
So the bootstrap really does not have traits.

Improving separation of the builder enhancers

Removing trait stuff from the kernel

Organizing Traits stuff

Only installing trait compositions if traits are available

Restoring the accessors to traitComposition

Fixing the initialization of Behavior

Creating the baseline

Fixing baseline

Packages with only extension methods are not supported by Ring2, adding the manifest to have a class.

Adding missing manifests, as Ring2 does not support extension only packages.

Fixing Hermes to detect if there is traits in the begining of the installation of the package.

During installation of a hermes package all the announcements should be stored and processed later

Implementing the method origin logic in the class, so it can be extended later.

Removing all the kernel stuff (partial)

Removing traits from the kernel (continuing)

Making the implementation more bootstrap friendly.