Now FFICompilerPlugin perform two operations in transform method:
1)  it collects all ffiCalloutTranslator methods
Because plugin processes all compiled methods we can do it just in transform method.
The selectors are collected in class variable FFICalloutSelectors which is initialized with known selectors.
It collects selectors of methods with pragma ffiCalloutTranslator.
 
2)  it marks all ffi methods as ffi. 
It checks that method sends one of the collected ffiCallout messages. And it adds special method property isFFIMethod.
And it forces ast to cache arg names in method property #argumentNames.
This approach removes  special kind of FFIMethodState which was proposed in original pull request. And it removes methodState methods like #decompiledArgumentAt:. Decompiler now simply checks for existing method property #argumentNames.

CompiledMethod>>argumentNames is now always check for value in properties and if #argumentNames is not exist it asks ast.

Senders of #ffiArgumentNames are now use  simple #argumentNames because with ffi plugin the value will be always cached.
I not remove #ffiArgumentNames because it used by generated ffi methods in my dev image. So remove will be another issue for cleanup.

This commit do not include possible protection which will return error when ffi method generated for method without isFFI mark. It can happens if new ffiCalloutTranslator method was introduced but it was not marked by pragma ffiCalloutTranslator