Terrible HACK!!!!
I am increasing the size of the eden space.
This allows to load the big baselines.
However, this is only needed because the VM has a bug when extending the space itself. It is corrupting the objects, this produces random crashes

I have detected that the problem is that the GC is corrupting the the objects when it has to enlarge the Eden.

This problem also was detected some months ago by Pavel when he tried to load a big baseline (Seaside in that moment).
I have solved it in a similar way, increasing the Eden size before loading BaselineOfIDE.