MirrorModificationForbidden exception is added as subclass of MirrorModificationForbidden. It overriddes #retryModificaiton method to use original mirror based selector.
In addition this fix makes MoridicationForbidden kind of Error instead of Exception.
Because many tools handle Error instead of general Exception (SUnit for example). And with previous version this failure was skipped by them which is wrong.
And generally this failure is kind of error.

Also better name for instance creation method:
	for: anObject at: fieldIndex with: newValue retrySelector:
	
And more tests to cover new behaviour