Make ScrollPane>>#scrollToShow: scroll as little as necessary
It also now takes into account making the rectangle visible in the horizontal direction.

This ensures that if any part of the rectangle was visible before, that part remains visible. Futher if only one of horizontal or vertical scrolling is needed to bring the rectangle in view the other position is left alone. That means that, for example, the tree in the profiler doesn't move all the way over to the right when moving through the tree with the keyboard but instead maintains the same horizontal position.

Also the horizontal scrollbar is properly updated to reflect the new offset.