When scrolling a tree node into view, show the node content not the gap
Previously, the node's bounds were made visible but this does not include the node content (e.g., icon or label). Added a method contentBounds to MorphTreeNodeMorph to give the bounds of the content, including the toggle and check, if any. This is then the rectangle that is made visible. You can easily see the effect on a deeply nested tree like that of the profiler.