21615-Separate-bootstrap-process-in-to-distinct-stages
The current bootstrap process has scope for improvement :-):

- Downloads occur at various stages during the process, which makes it
  difficult to test new code in a personal environment, e.g. choosing a
  specific VM to bootstrap with.
- Code is duplicated in Jenkinsfile and the various shell scripts
- It isn't easy to snapshot the process and resume bootstrapping from
  that point (related to the first item).

The proposed patch is a first step that will restructure the shell
scripts in to the following phases:

1. Clean up (remove any artefacts from previous runs)
2. Download all required files
3. Prepare the image for bootstrap
4. Build the new image

This work is intended to complement the existing effort by Pablo and
others to move more of the code from shell scripts to Pharo.

A second PR will refactor Jenkinsfile to call the new scripts, removing
the code duplication.