Remove gradient management from PluggableButtonMorph.
The management of the gradient from PluggableButtonMorph works only in the theme define gradient fill style and just mess with the theme when it's not the case.

I think the gradient should be managed via the theme only and not via the button AND the theme. So I remove the part from the button to only use the theme one. 

This correct https://pharo.fogbugz.com/f/cases/21928/Remove-PluggableButtonMorph-option-to-use-gradient-since-it-s-not-used-and-style-it-decided-by-the-theme but also https://pharo.fogbugz.com/f/cases/21894/PluggableButtonMorph-does-not-reset-its-mouse-over-fill-style-after-the-mouse-leave-the-morph