22504 OSPlatform>>currentWorkingDirectoryPathWithBuffer: fails
OSPlatform>>currentWorkingDirectoryPathWithBuffer: has a number of issues:

1. It doesn't find the primitive because the module isn't specified.
2. The primitive call fails because the primitive expects 0 arguments while the method passes one (which is unrelated to the primitive).

I'm not sure why the FFI fallback is required.  If the primitive is failing, it should be fixed (I don't see how the primitive can fail under normal circumstances, it is effectively just getcwd()).  Replacing the primitive with the FFI call is problematic at the moment as it fails sometimes, depending on which VM is used.

For now I'll just fix the code leaving the functionality as intended, and we can revisit it later if desired.