Adding a superclass to the Glamour tests to handle the waiting for UI redrawn in a more tidy way.
Glamour Tests are using directly World >> #doOneCycle.
This produces a problem when executing in the CI.
As the tests in the CI are run in another process different than the UI process, if we do #doOneCycle there is the risk of race conditions.

This will improve the stability of the build in the CI.