Restore #getEnv:
that used to be public API

The new impl handles the characters outside ASCII. For the strings that were valid before the new impl will return exactly the same. The change is for the environment values using characters with codepoint > 127. For example accented leters or japanese characters
We got to this, because the launcher is used by users with usernames in french, spanish or japanese