Remjove Stream and Collection  #write: (as an alias to #<< and #add: respectively)
This selector is not used for these 2 receiver types in the standard image (except for one unit test), I never saw it being used in actual code.

These are just aliases that add nothing but cognitive load.

They make the API's of two important hierarchies (and especially the interface for non-inheriting classes) heavier than needed for no functional benefit.

https://github.com/pharo-project/pharo/issues/2701