Convert the stack pointer from alien to ExternalAddress early
This way >>#handle:at: will work with the ExternalAddress instead of an alien and FFIExternalStructure handle (or handle of handles) will work with an address. In case of structs (and not pointers to struct) this means they reference the stack and are invalid once the FFI call returns. While this looks scary it seems to be desired over a copy.