Make GTSpecPreDebugger work with Spec 2.
WARNING: This will break the possibility to open the full debugger by clicking on the stack for now. This will be fixed by the next Spec2 integration. The full debugger can still be opened by clicking on the "Debug" button. 
I propose the change because it'll make it easier to work on Spec2 like this (currently we get an emergency debugger at each error...) and will already be done for the next Spec2 integration.

Fixes #3170