In the compiler, we have a dedicated subclass of Set to check identity using #literalEquals.
This actually can be replaced by a PluggableSet configured correctly to use #literalEquals:

PluggableSet new equalBlock: [ :a :b | a literalEqual: b]

fixes #4020 
fixes #4021

(Compiler tests are green).

Why is this good? It removes on subclass of Set. Subclassing basic collection classes can be sub-optimal as they assume too much about the implementation details of the superclass.

Note: as #literalEquals: is a Kernel method, we could add a class side instance creation method to PluggableSet in the style of #integerSet