Fixes 2936 The test for isSelector is not done in the test class for RBScanner, but there is an extensive suite of tests in RBConditionTest (Should I raise an issue to have them moved???). I reincluded the two outcommented tests in testCheckInvalidMethodName
My brief view on this bug is that selectors are *not* part of the smalltalk grammar, and as such there was no good mechanism and tokens to use. The current version failed for that reason, trying to get by using what was there. I added a new method (and its helper), to check for validity of selectors. It is a small state machine, see documentation in helper class.