For the FileDialog of Pharo 8 I want to use an undo/redo system. I checked HistoryIterator but I find it too complicated for my needs.
In this PR I introduce a new history iterator. You configure it with an undo and redo block and then you can directly register objects you navigate into. When and undo or redo will happen, the objects will be used as parameter of the valuables to execute the action.

For example check test class and class comment of ConfigurableHistoryIterator.