Fixes issue 3068
It seems that the is a near solution already in the discussion of the issue. 
There is a method #firstComment, which should return the first comment. 
I changed the method to actually return an ordered collection of comments in the method.
I changed the test method to depend less on the comments in other parts of the system, and noted where it does.