Fixes: #2372
I added a "scanTokens:" to RBScanner. I do not believe there should be a helper method in RBParser as the parser should focus on AST, but I am still new to all this.
The method was nearly there. I build on the #contents method of the scanner. Since the issue mentions a similarity to the #split method, I extract the token values from the scanner, and ignores the token types.
I added a test method as well, which check that white space and comments are removed.