Fixes #4365
I added the tests which was implicitly part of the original issue.
Besides Symbol, no other subclass of SequenceableCollection (which is the least superclass of SortedCollection to implement #,.).

The issue is only for sorted collection, as is a simple case of not copying the instance variable in sorted collection to the new object.