Add new #printOneLineString and related methods
It's nice to have a #printOneLineString that is just like #printString
but avoids printing CRs and LFs. This patch implements such idea adding
also a method named #printOnOneLine: which will get called with a
FilteringStream implementing the required behaviour.

The FilteringStream is used to replace CRs and LFs with a space, but can
also be used to filter elements going into a Stream via a generic block.

Fixes: #4849