Replace #printString implementations with #printOn:
On MalGraphEdge, MalHitsNode and MalWeightedEdge, the #printString
implementation was calling #printString on their ivars, causing the
creation of too many temporary Streams and spurious copies.

This patch replace those methods, when needed, with a #printOn:
implementation. In that way, we avoid creating substreams and spurious
copies.

Some unit tests have been added, and they were running also before the
implementation change, this is to ensure that the former behavior have
not been changed.

Fixes: #4855

Co-authored-by: Vincent Blondeau <vincent.blondeau@polytech-lille.net>