Simplify Parsing in the case of Syntax Hightlighting.
The method #parse:isMethod:  used to have an "on: Error" handler. Which actually got called because the message send used to be a DNU
(becuause #parseFaultyMethod: was returning sometimes RBParseErrorNode, not a method node...).

As we fixed #parseFaultyMethod: to always return a method, we  can now do this  much cleaner.