This is fixing a bug while loading code.
In Pharo 7, the loading of code is using a cache for MCDefinitions. 
When you create a new definition, it checks if one equivalent was already created, if yes, it return this one.

The equality of MCClassTraitDefinition does not take the category into account and I got bug while loading code because the MCClassTraitDefinition was wrong because it found one with another category.