- use the pragma collector to just find all the selctors tagged with #ffiCalloutTranslator - remove the magic to update if methods with that pragma are added (as that is the reason why we have the problem)
If people now add new ffi callout selectors, they have to do "FFICompilerPlugin initialize". Which is ok as I think this is a bad idea anyway to do that.

This needs another change in the  #postLoadActions of  LibtGit:

	FFICompilerPlugin initialize.
	
did the change in my image but it does nor apear in the PR. Someone else has to do that.