GT Inspector: Fix Hook for Unary Presentation Generator
This code is (probably only) used by Magritte to declaratively create GT Inspector presentations. I (@SeanDeNigris) [originally introduced it](https://github.com/moosetechnology/Moose/issues/1130), but it seems to have been altered (I'm assuming accidentally, possibly as a result of a system-wide method rename, but can't tell for sure because the change seems to predate the move to GH).