Next step in the methodMap cleanup: remove the method related to "namedTempAt:".
the debugger methodmap has two kinds of offsets for temps: the real one (tempAt:) and a "virtual" offset for all real temps that are accessible by name.
This makes the whole thing quite difficult to understand and is better removed. There are no senders in the image, the methods where private to the debugger.