The OCTempVariables used to cache the index after the first access. The index itself was taken from the IR.
We can do it differently: every time a temp is created on the IR level, we can backfill the offset in the OCTempVariables.

This PR:

- IRBuilder now fills the index of semantic variables on #addTemp:
- This allows to not search for the offsets in the IR when accesing temps reflectively (#readFromLocalContext: and similar)

Fixes:
- some improvements to #lookupVar: (so we can call it with non-string var names)
- map the node correctly when generating inlined blocks and full blocks

Cleanups:
- remove #tempVectorNamed: #indexFromIRForVarNamed: #indexFromIR and #indexInTempVectorFromIR:

Accessing the index of OCTempVariables continues to do a nil check. This allows us to call it even if the IR has not yet
been generated.