SyntaxErrorDebugger, if we close it, messes up the state of the morphic rendering.
The reason is that it suspends the morphic process and creates a debugger, but when closing it does not terminate the debugger so the rendering is left in a bad state.

It used to be that #release was called to do that, but it is not anymore.

Fix: call it in #closeWindow.

In addition, this PR contains two trivial cleanups. fixes #6351