We do not need the #doesNotUnderstand: Handler on the CompilationContext.
The ideas was to have some ultra dynamic way that options can be set and then reacted on by plugins without the need to add code.

But for that we do not need the DNU: plugins just have to query the options using the API instead of just calling #optionSomething.