At some point we added a tool to show compiler scopes... but in the end it is not that useful
In addition, we need to merge the scope for global with "Smalltalk globals" and the class scope with the class. This means that we will not instantiate new objects for each method but reuse the globals scopes, making a general "children" concept impossible.
In addition, this is a bit wasteful considerung that just one specialised tool was using the information.

Therefore: this PR removes the Scopes inspector and cleans up the code of OCAbstractScope.

If we find out we miss the tool really a lot, we can bring it back just for instance scope and downwards from there easily.