ClassVariable was there as a sublass of ClassVariable, with the idea to implement queries (e.g. #usingMethods) there.
But this is not a good idea, as it would mean that we could not use any sublasses (e.g. WeakClassVariable) as pools.

We therefore remove the class (it was not used) and yes, will have to deal with queries in the methods of ClassVariable to treat pool vars specially  (as we do now)