- FIx classbuilder:  move #installSlotsAndVariables to before recompiling methods - add UndefinedSlot
The idea of UndefinedSlot is that for the cast a class is loaded where the slot definition does not exist, we create this special UndefinedSlot.
-  It has the same name
- has the definition of the slot (the AST) 
- prints itself as the definition (the AST). This might be slightly different then the prefered printout for the slot, but contains all infos to create the correct slot later
- reads return nil
- writes are ignored

On read and write, the UndefindedSlot will check if the class of the original definition now has been loaded and if yes, will rebuild the class.