This speeds up class deletion by not putting deleted classes into the Obsoletes
As we did not recompile, existing references still referenced the obsolete class anyway. Just the syntax coloring was hiding that.

Thus, it is better to make that explicit by adding a Rule that shows when a method references obsolete classes.

small additional cleanup: ReGlobalVariablesUsageRule uses now #isGlobalClassNameBinding instead of having a reference to the Smalltalk globals.