This adds a nameIsReserved: check to ReservedVariable plus a test.
The idea is to use that in the compiler and e.g. when we check variable name validity. If we do that, we can add new Reserved Variables by just adding a subclass.

(this is not as simple to do to use it in Slot class as it has implications for dependencies, ReservedVariable is for now in the compiler package,
therefore this PR does not do that change yet. The Compiler change will be done im combination with removing all hard-coded references to the variable names)