!!!! Highly Experimental!!!
This is disabled by default.

If you want to experiment, you can enable the #optionCleanBlockClosure compiler option.

What does it do? For blocks that do not access anything outside (#isClean is true), it statically creates a FullBlockClosure and pushes it with #pushLiteral.

This avoids having to create the FullBlockClosure from the CompiledBlock at runtime, which is much faster.

This is not yet complete, see the class CleanBlockClosure for more comments and TODOs.