Make sure Self/Super/ThisContext node have the right variable.
In the new version of the AST, VariableNodes needs a reified variable. In case of Self/Super/ThisContext it exists a singleton for it. 
Instead of setting it by hand, those nodes are now initialized with their variables. This fix bugs where AST could have variable nodes without any variable attribute.

Fixes #7077