This is a first idea how to cache pragmas -> Pragma class has a Dictionary with one entry per selector -> It references the Pragmas weakly
CompiledMethods add their pragmas to the cache when they are added in a class.

Speed to get all pragmas:

[Pragma allInstalled] bench "'799.320 per second'".
[PragmaCollector allSystemPragmas] bench  "'54.923 per second'"

But it will now grow with the number of pragmas, not number of methods.

Pragma allNamed: aSymbol can replace the use of the PragmaCollector in many cases (to be explored)

This PR just adds the Cache and tests, it is not yet used. The next step is to test it a bit, compare it with #allSystemPragmas and
the implement allSystemPragmas to use the cache.