Historically, sending #binding to a Class gives you the association in the systemDictionary.
We have now changed this so that those objects are the Variables representing the concept of accessing the global variable.

That is, 

OrderedCollection variable

now returns the *Variable* that references OrderedCollection.

To be in sync everywhere, we should use the selector #variable to get Variables. We did that already on the AST node level, this PR does the same change on the level of Behaviors.

-> the implemetation is now #variable
-> binding is retained as a compatibility method.

A later PR then can change all users of binding to use #variable instead