Lots of methods and blocks refer to literals that are equal.
This is not yet turned on by default as literals can be changed... which leads to bugs that are now local, but with this change get to have a potential global effect. (making exactly those literals read only that are shared here is a future step).


ImageCleaner new statistics "141930" 

ImageCleaner shareLiterals

ImageCleaner new statistics "118979"