EpMonitor: Remove old code for subscribing to test announcer.
By 2015, Epicea logged test runs and results. For example, when Test Runner run a suite, the log registered the successes and failures. At some point, that became a package extension, which is still in SmalltalkHub as "Epicea-SUnit". That package requires this code in EpMonitor that I propose to delete in this issue. Why? 1. it hasn't been used for years but it'd be easy to recover after delete; 2. AFAIR, the feedback I got was that logging test runs implied too much noise in the log of changes.

Fixes #7379.