The classes modeling self, super and thisContext reference the global environment directly (via SystemNavigation new).
This PR instead adds the environment as state. This will allow us to later have different instances for different environments. For now the envrionment remains hardcodes, but just in #initialize instead of all #usingMethods implementations.

(The idea is that all subclasses of Variable should implement #scope. They should know the scope that will return them when looking up by name using #lookupVar:. The scope for self, super and thisContext will be the global environment)

The PR in addition adds a missing visitor methid to the completion builder.