#typeOfClass is a huge "IF" that actually misses cases: DoubleWordArray and DoubleByteArray even if the layout is correct will answer wrongly.
Solution: forward to the layout, implement #typeOfClass there. Then it will for DoubleByteArray automatically return the default case, the name of the layout class (#DoubleByteLayout), which should make it then possible to save classes like that with Git.

we can remove #isCustomLayout now, too. Custom layout now use the name of the layout class by default.