Imagine you define a binding in a workspace:
temp := 1.

and then you "doit" this code:

self halt

you expect the variable to be shown in the debugger as a temp. 

The PR is a fix to do exactly that.
- we need to make sure to add the doit compiled method to the AST cache
- #allTemps of the Requesterscope needs to return a list of all Variables defined by the Requestor

And it works! it seems there are sometimes odd bindings for "self" generated, we need to check why but that is another PR