referredInstVars has no senders
everything that you might want to do with this is much better done on the level of the First Class Variables that we have now.

referredInstanceVariables
 ^self methodClass  instanceVariables select: [ :each | each usingMethods includes:  self ]

Which then takes into account all kinds of self defined ivars... and returns *objects*, not strings.

Thus we can add a method for this as soon as we have any need...I am sceptical if we really need it, though.