Class references show incorrect results when there are super sends in blocks.
An example can be seen in DelayMutexScheduler>>#unschedule:, if you lool for references to DelayMutexScheduler, it shows this method.

The reason is that "super" in full block is compiled as a push of a binding to the class and then a directed send. We push just "theClass binding".
This will create a normal literal for that binding, while in a method it will find the last literal that is ignored for literals as it is always the class binding.

Solution: push a copy.

Fixes #8701