add #resetReservedVariables
The reason is that we want to be able to add new kinds of Variables like self, thisContext as experiments.

For this we can easily add a subclass to ReserverdVariable, but we do need to make "Smalltalk globals reservedVariables" take them into account.
This can now be done by calling #resetReservedVariables. We could call this even in the name analyis by default once (need to benchmark).