Read and Write of indexed instance variables via the variable (IndexedSlot or a subclass) is using the mirror method on thisContext. (thus we can read ivars from objects without sending a message to them, Variables thus can be understood to be special kinds of Mirror objects)
```Smalltalk
read: anObject
	^ thisContext object: anObject instVarAt: index
```

But this means that we have to reference thisContext... which the VM does not like (it is slow).

The actual primitives do not care about the receiver, we can just copy them and implement them in IndexSlot:

```Smalltalk
read: anObject
	^ self object: anObject instVarAt: index
```

A simple naive benchmark:

```
slot := Point slotNamed: #x.
[slot read: 4@4 ] bench  

before: "'12863984.406 per second'"
after: "'28472842.495 per second'"
```

Which means this is indeed faster, with not changing the complexity too much.