We should think about Shadowed Variables on the level of the class
Classes define instance variables (modeled by subclasses of Slot) and Class Variables. They can shadow Variables from the environment
that the class is in (normally this is Smalltalk Globals).

We should extend the concept of Shadowed variables to over these cases, too. 

Imagine we have a class that has a class variable "Person". Then we add a class named "Person": The class variable is now shadowing
the class. #isShadowing is just checking for the property. But here we run into a problem: who could set the property? The classbuilder
could do it. But adding a class then would mean that we would need to chck all existing classes if the new global is shadowed somewhere.

And we would not just have to check the classes: nothing forbids us to use a block argument "Person"... thus we would need to re-analyse
all code globally. Not a good idea in a system where current users might have >100K classes...

Thus: we need to rethink our solution. It was not wrong: we were concened with just modeling shadowing for methods, the solution to tag the
variables was a good one fort that use. But now we need to step back. What is shadowing again? It means that the outer scope of a Variable has 
already a variable of the same name. 

Interestingly, the Variable model is already modeling all the needed information. We can just change  #isShadowing to do that:

```Smalltalk
isShadowing
	"I shadow a variable if looking up my name in the outer scope returns a variable"
	| outer |
	outer := self scope outerScope.
	^(outer ifNotNil: [self scope outerScope lookupVar: self name]) notNil
```

We just have to add a missing #outerScope method to Behavior (as the outer scope is called the environment here). And it works! The tests
we did for shadowing are still green.